* CorApi2\Metadata\CorMetadata.cs: * CorApi2\Metadata\MetadataType.cs: The Make*Type methods are now   public.
* MonoDevelop.Debugger.Win32\CorEvaluationContext.cs: New helper
  property.

* MonoDevelop.Debugger.Win32\CorObjectAdaptor.cs: Improved
  CallToString. Fixed enum conversions. Implemented some missing
  overrideables: CreateTypeObject, GetEnumMembers,
  GetIndexerReference.

* MonoDevelop.Debugger.Win32\CorDebuggerSession.cs: Improved
  implementation of GetTypeInfo. It now supports primitive, generic
  and array types.

* MonoDevelop.Debugger.Win32\FieldReference.cs: When getting enum
  members, convert the integer value to an enum value

* MonoDevelop.Debugger.Win32\PropertyReference.cs: Improved
  implementation of indexers. Properly box primitive values when
  setting/getting the property.

* MonoDevelop.Debugger.Win32\MonoDevelop.Debugger.Win32.csproj:
  Updated.

svn path=/trunk/monodevelop/; revision=154864

+++ extras/MonoDevelop.Debugger.Win32/CorApi2/Metadata/CorMetadata.cs
@@ -684,13 +684,13 @@ namespace Microsoft.Samples.Debugging.CorMetadata
 						Type t = ReadType (importer, ref pData);
 						uint rank = CorSigUncompressData (ref pData);
 						uint numSizes = CorSigUncompressData (ref pData);
-						List<uint> sizes = new List<uint> ();
+						List<int> sizes = new List<int> ();
 						for (int n = 0; n < numSizes; n++)
-							sizes.Add (CorSigUncompressData (ref pData));
+							sizes.Add ((int)CorSigUncompressData (ref pData));
 						uint numLoBounds = CorSigUncompressData (ref pData);
-						List<uint> loBounds = new List<uint> ();
+						List<int> loBounds = new List<int> ();
 						for (int n = 0; n < numLoBounds; n++)
-							loBounds.Add (CorSigUncompressData (ref pData));
+							loBounds.Add ((int)CorSigUncompressData (ref pData));
 						return MetadataType.MakeArray (t, sizes, loBounds);
 					}
